% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memprof.R
\name{monitor_from_log}
\alias{monitor_from_log}
\title{Build \code{memprof_monitor} object from log file which can be used for plotting.}
\usage{
monitor_from_log(log_file)
}
\arguments{
\item{log_file}{Path to memprof log file.}
}
\value{
A \code{memprof_monitor} object.
}
\description{
Build \code{memprof_monitor} object from log file which can be used for plotting.
}
