% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memprof.R
\name{monitor_read}
\alias{monitor_read}
\title{Read monitor log file as a \code{memprof_use} data frame for plotting.}
\usage{
monitor_read(monitor_file)
}
\arguments{
\item{monitor_file}{Path to memprof monitor log file.}
}
\value{
A \code{memprof_use} data frame.
}
\description{
Read monitor log file as a \code{memprof_use} data frame for plotting.
}
