% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memprof.R
\name{plot.memprof_result}
\alias{plot.memprof_result}
\title{Plot memprof monitor result}
\usage{
\method{plot}{memprof_result}(x, ...)
}
\arguments{
\item{x}{The \code{memprof_result} object}
}
\value{
Nothing, creates a plot.
}
\description{
Plot memprof monitor result
}
