% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memprof.R
\name{with_monitor}
\alias{with_monitor}
\title{Run an R expression and monitor memory}
\usage{
with_monitor(
  expr,
  poll_interval = 0.1,
  monitor_file = tempfile(),
  overwrite = FALSE,
  gc = TRUE
)
}
\arguments{
\item{expr}{The expression to evaluate.}

\item{poll_interval}{Time between each check of memory use.}

\item{monitor_file}{Path to log memory monitor to, you can use this if
your expr errors to recover the output for later plotting. Uses a
tempfile by default.}

\item{overwrite}{If TRUE then overwrite \code{monitor_file} if it exists already.}

\item{gc}{If \code{TRUE} run a garbage collect before starting monitoring.}
}
\value{
A \code{memprof_monitor} object containing the \code{result} as the outcome
from the expression and \code{memory_use} the memory used
}
\description{
Evaluate an expression, and monitor the memory every \code{poll_interval}.
This will monitor the memory of your entire system, not just the expression
being run.
}
